An old and hidden printing curse was lurking in this place for all these years: - The functions SERVER_Printf and SERVER_PrintfPlayer were missing the GCCPRINTF macro, needed for non-MSVC compilers to do format checking for their calls. Noticed when Dusk's votedef code crashed while calling a SERVER_PrintfPlayer (provided 1 string parameter in excess), even though the compiler said it was fine. - Thanks to that, fixed all the new compiler warnings which appeared afterwards (essentially, LONG arguments for '%d', no format string at all for single string parameters, etc). While I'm at it, changed the name of the function SERVER_RCON_Printf to SERVER_RCON_Print, because it does not do format printing at all.

+++ src/d_main.cpp
@@ -355,7 +355,7 @@ CUSTOM_CVAR (Int, dmflags, 0, CVAR_SERVERINFO | CVAR_CAMPAIGNLOCK)
 	// [BC] If we're the server, tell clients that the dmflags changed.
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( gamestate != GS_STARTUP ))
 	{
-		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (LONG)self );
+		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (int)self );
 		SERVERCOMMANDS_SetGameDMFlags( );
 	}
 }
@@ -411,7 +411,7 @@ CUSTOM_CVAR (Int, dmflags2, 0, CVAR_SERVERINFO | CVAR_CAMPAIGNLOCK)
 	// [BC] If we're the server, tell clients that the dmflags changed.
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( gamestate != GS_STARTUP ))
 	{
-		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (LONG)self );
+		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (int)self );
 		SERVERCOMMANDS_SetGameDMFlags( );
 	}
 
@@ -509,7 +509,7 @@ CUSTOM_CVAR (Int, dmflags3, 0, CVAR_SERVERINFO)
 	// [BB] If we're the server, tell clients that the dmflags changed.
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( gamestate != GS_STARTUP ))
 	{
-		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (LONG)self );
+		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (int)self );
 		SERVERCOMMANDS_SetGameDMFlags( );
 	}
 }
@@ -546,7 +546,7 @@ CUSTOM_CVAR (Int, compatflags, 0, CVAR_SERVERINFO)
 	// [BC] If we're the server, tell clients that the dmflags changed.
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( gamestate != GS_STARTUP ))
 	{
-		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (LONG)self );
+		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (int)self );
 		SERVERCOMMANDS_SetGameDMFlags( );
 	}
 }
@@ -562,7 +562,7 @@ CUSTOM_CVAR (Int, compatflags2, 0, CVAR_SERVERINFO)
 	// [BC] If we're the server, tell clients that the dmflags changed.
 	if (( NETWORK_GetState( ) == NETSTATE_SERVER ) && ( gamestate != GS_STARTUP ))
 	{
-		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (LONG)self );
+		SERVER_Printf( PRINT_HIGH, "%s changed to: %d\n", self.GetName( ), (int)self );
 		SERVERCOMMANDS_SetGameDMFlags( );
 	}
 }
